app/requirements.txt
Flask[async]
flask-sqlalchemy
psycopg
psycopg2-binary
gunicorn